Add files via upload
The Tip Calculator is a simple web application that allows users to calculate the tip and total amount based on their bill. Users can input the bill amount, tip percentage, and the number of people to split the bill. The app supports both Indian Rupees (INR) and US Dollars (USD), allowing users to choose their preferred currency. It then calculates the tip amount, total amount, and the amount each person needs to pay, providing an easy way to manage payments after a meal.